refactor: move entities in fields
