+++ apps/film/src/entities/genre/film_genre.entity.ts
@@ -5,7 +5,7 @@ import {
   Model,
   Table,
 } from 'sequelize-typescript';
-import { Film } from './film.entity';
+import { Film } from '../film/film.entity';
 import { Genre } from './genre.entity';
 
 @Table({ tableName: 'film_genre', timestamps: false })
@@ -22,6 +22,6 @@ export class FilmGenre extends Model<FilmGenre> {
   film_id: string;
 
   @ForeignKey(() => Genre)
-  @Column({ type: DataType.INTEGER })
+  @Column({ type: DataType.STRING })
   genre_id: string;
 }
