+++ apps/film/src/entities/index.ts
@@ -1,7 +1,9 @@
-export * from './film.entity';
-export * from './genre.entity';
-export * from './film_genre.entity';
-export * from './trailer.entity';
-export * from './language.entity';
-export * from './film_language_audio.entity';
-export * from './film_language_subtitle.entity';
+export * from './film/film.entity';
+export * from './quality/quality.entity';
+export * from './quality/film_quality.entity';
+export * from './genre/genre.entity';
+export * from './genre/film_genre.entity';
+export * from './trailer/trailer.entity';
+export * from './language/language.entity';
+export * from './language/film_language_audio.entity';
+export * from './language/film_language_subtitle.entity';
