+++ apps/film/tsconfig.app.json
@@ -4,6 +4,6 @@
     "declaration": false,
     "outDir": "../../dist/apps/film"
   },
-  "include": ["src/**/*"],
+  "include": ["src/**/**/*"],
   "exclude": ["node_modules", "dist", "test", "**/*spec.ts"]
 }
